requirements.txt
# Core LangChain and LangGraph
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0

# Vector Store and Embeddings
langchain-chroma>=0.1.0
langchain-huggingface>=0.1.0
langchain-ollama>=0.2.0
chromadb>=0.5.0
sentence-transformers>=2.0.0

# LLM Integration
langchain-openai>=0.2.0

# Utilities
python-dotenv>=1.0.0
chardet>=5.0.0

# Web Scraping (for crawler)
playwright>=1.40.0
beautifulsoup4>=4.12.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
